refactor: comment 반영 - quiz -> test 로 명명법 수정 및 관심사 분리
